# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Instay backend API documentation",
        "title": "Instay API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/files/presigned-urls/uploads": {
            "post": {
                "description": "Tạo một hoặc nhiều URL (presigned) để client upload file lên storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get Upload Presigned URLs",
                "parameters": [
                    {
                        "description": "Danh sách file cần tạo URL (tên file, content type)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UploadPresignedURLsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tạo URL upload thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "presigned_urls": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/types.UploadPresignedURLResponse"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (validation error)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/files/presigned-urls/views": {
            "post": {
                "description": "Tạo một hoặc nhiều URL (presigned) để client xem/tải file từ storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get View Presigned URLs",
                "parameters": [
                    {
                        "description": "Danh sách file key cần xem",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ViewPresignedURLsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tạo URL xem thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "presigned_url": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/types.ViewPresignedURLResponse"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (validation error)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "types.UploadPresignedURLRequest": {
            "type": "object",
            "required": [
                "content_type",
                "file_name"
            ],
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                }
            }
        },
        "types.UploadPresignedURLResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "types.UploadPresignedURLsRequest": {
            "type": "object",
            "required": [
                "files"
            ],
            "properties": {
                "files": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/types.UploadPresignedURLRequest"
                    }
                }
            }
        },
...
        "types.ViewPresignedURLResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "types.ViewPresignedURLsRequest": {
            "type": "object",
            "required": [
                "keys"
            ],
            "properties": {
                "keys": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}